Create Program.cs
This program generates data for modified Bingo cards. Modified means that every card contains the same manadory numbers. We once needed this for a sketch in a show where everone in the audience wins. The generated textfile can be imported into an Excel file that defines the layout.

aantalKaarten contains the number of cards to generate.
verplichteGetallen contains a list of the mandatory numbers per card.