Bingo Cards template
Kienkaarten.xlsx contains a template for the Bingo Cards. How it works:
Open the generated text file and copy all to clipboard.
in the Excel file, open a new sheet and paste the clipboard there. Convert the semicolon separated lines into their cells. Copy all and paste it text only into cell A1 of the template sheet.
I know it is somewhat of a hassle, but there was no need for a fancy user interface and fulfilled its purpose just fine.